# server/openapi.yaml
...
openapi: 3.1.0
info:
  title: IMS REST API
  version: 1.0.0
  description: >-
    Web-Based Inventory Management System API. Contract per REST-API
    Specification 05 (approved): additive evolution only within v1; clients
    MUST tolerate unknown response fields. All routes relative to /api/v1
    except GET /health and GET /ready (server root, ahead of limiters/auth).
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/google:
    post:
      tags: [auth]
      operationId: googleAuth
      summary: Sign in / sign up with Google (GIS ID-token → session)
      description: >-
        Verifies a Google Identity Services ID token (google-auth-library) and
        resolves-or-provisions by VERIFIED email: an existing account is linked
        (googleSub attached) and logged in; an unknown email provisions a new
        password-less workspace. Backs both the "Sign in" and "Sign up with
        Google" buttons. Unverified Google email is rejected (401).
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
      responses:
        '200':
          description: >-
            Session established. Also sets the rotating refresh cookie
            (httpOnly · Secure · SameSite=Strict · Path=/api/v1/auth).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '400':
          description: >-
            Missing credential (VALIDATION_ERROR) or Google sign-in not
            configured on this server.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Invalid Google token, or the Google account email is unverified.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '409':
          description: DUPLICATE_EMAIL — a provisioning race on the same email.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          $ref: '#/components/responses/RateLimited'
...
  /auth/reset-password:
    post:
      tags: [auth]
      operationId: resetPassword
      summary: Complete an Admin-initiated reset (FR-AUTH-05, UC-03)
      description: >-
        Token is single-use, 30-min expiry, hashed at rest; success revokes
        all prior sessions and clears mustChangePassword.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '204':
          description: Password installed; all prior sessions revoked.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Invalid/expired/used token (UNAUTHORIZED).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '429':
          $ref: '#/components/responses/RateLimited'
  /auth/change-password:
    post:
      tags: [auth]
      operationId: changePassword
      summary: Change own password (FR-USER-05)
      description: >-
        Current password required and verified; revokes all OTHER sessions.
        One of the three calls a mustChangePassword session may make (AAD §2).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
      responses:
        '204':
          description: Password changed; other sessions revoked.
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          description: Wrong current password, or missing/invalid access token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: refreshToken
      description: >-
        Rotating refresh session — httpOnly · Secure · SameSite=Strict,
        scoped to /api/v1/auth (AAD §3.2). Never appears in bodies or headers.
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >-
        Access token — 15-min TTL, payload sub/role/iat/exp only (SEC-01).
        The rotating refresh token travels exclusively as an httpOnly cookie
        scoped to /api/v1/auth and never appears in this contract's bodies.
...
  responses:
    ValidationError:
      description: Request failed schema validation before the controller ran (05 §1).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
    RateLimited:
      description: Rate limit exceeded (RATE_LIMITED, SEC-04).
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
...
  schemas:
    ErrorCode:
      type: string
      description: >-
        The CLOSED set of codes this API may emit — SRS §16.3 catalog plus the
        three documented APR-01 extensions (DUPLICATE_EMAIL, PRODUCT_NOT_EMPTY,
        PRODUCT_HAS_HISTORY). 05 §6.2.
      enum:
        - VALIDATION_ERROR
        - UNAUTHORIZED
        - ACCOUNT_DEACTIVATED
        - FORBIDDEN
        - NOT_FOUND
        - DUPLICATE_SKU
        - DUPLICATE_BARCODE
        - DUPLICATE_EMAIL
        - INSUFFICIENT_STOCK
        - STALE_WRITE
        - LAST_ADMIN
        - PRODUCT_ARCHIVED
        - PRODUCT_NOT_EMPTY
        - PRODUCT_HAS_HISTORY
        - CATEGORY_IN_USE
        - INVALID_BARCODE
        - IDEMPOTENCY_CONFLICT
        - ACCOUNT_LOCKED
        - RATE_LIMITED
        - INTERNAL_ERROR
        - SERVICE_UNAVAILABLE
...
    ErrorEnvelope:
      type: object
      description: The single error envelope every non-2xx JSON response uses (SRS §16.2, 05 §6.1).
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, correlationId]
          properties:
            code:
              $ref: '#/components/schemas/ErrorCode'
            message:
              type: string
              description: Human-readable; the frontend maps code → message via its own table (05 §6.2).
            details:
              description: Optional, code-specific payload (e.g. available/requested for INSUFFICIENT_STOCK).
            correlationId:
              type: string
              description: Echoes X-Correlation-Id — surfaced in error toasts (NFR-23).
...
    Role:
      type: string
      description: The two operator roles (SRS §5 matrix).
      enum: [ADMIN, STAFF]
...
    GoogleAuthRequest:
      type: object
      description: Google Identity Services ID token; verified server-side.
      required: [idToken]
      properties:
        idToken:
          type: string
          minLength: 1
    ResetPasswordRequest:
      type: object
      description: 05 §15.7 — BR-32 policy applies to newPassword.
      required: [token, newPassword]
      properties:
        token:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 10
          maxLength: 64
    ChangePasswordRequest:
      type: object
      description: 05 §15.7 — new ≠ current; BR-32 policy on newPassword.
      required: [currentPassword, newPassword]
      properties:
        currentPassword:
          type: string
          minLength: 1
        newPassword:
          type: string
          minLength: 10
          maxLength: 64
...
    SessionUser:
      type: object
      description: The 05 §7.1 session user block — never carries credential fields (SEC-02).
      required: [id, name, email, role, mustChangePassword]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
          format: email
        role:
          $ref: '#/components/schemas/Role'
        mustChangePassword:
          type: boolean
    SessionSettings:
      type: object
      description: >-
        FCM-01 (RATIFIED 2026-07-23): read-only display constants in the
        session payload — Staff has no other approved endpoint for them
        (GET /settings stays Admin-only). Additive, EXT-01-compliant;
        exposes no settings management. AAD §12 · SMA §5.
      required: [systemCurrency, movementWarningThreshold]
      properties:
        systemCurrency:
          type: string
          description: ISO 4217 code — feeds lib/formatters.
        movementWarningThreshold:
          type: integer
          minimum: 1
          description: Large-movement confirm threshold (BR-15) for the movement dialogs.
    SessionResponse:
      type: object
      description: Login/refresh 200 — access token + profile + FCM-01 display constants.
      required: [accessToken, user, settings]
      properties:
        accessToken:
          type: string
          description: JWT HS256, 15-min TTL, payload sub/role/iat/exp only (SEC-01).
        user:
          $ref: '#/components/schemas/SessionUser'
        settings:
          $ref: '#/components/schemas/SessionSettings'
...